# Conflicts:
#	space-core/src/main/kotlin/com/ideaspace/core/Utils.kt
#	space-core/src/main/kotlin/com/ideaspace/core/repository/DocumentRepository.kt
#	space-core/src/main/kotlin/com/ideaspace/core/repositoryImpl/CrudDocumentRepositoryImpl.kt
#	space-sync-server/src/main/kotlin/com/ideaspace/document/Routes.kt

+++ space-sync-server/src/main/kotlin/com/ideaspace/workers/KafkaPartitionProcessor.kt
@@ -48,7 +48,7 @@ class KafkaPartitionProcessor(
             println("process record: ${record.value()}")
             if(messageCounter[key] == BATCH_LIMIT){
                     println("IO CONTEXT - Process flush kafka message to database")
-                    documentRepository.updateOffset(key, record.offset().toInt())
+                    documentRepository.updateOffset(key, record.offset())
                 messageCounter[key] = 0
             }
         }
